initial commit
Start port of muon.r MAE functions to rhdf5.
Includes a workaround for missing features in rhdf5. The workaround
works for file creation properties, but not for references, because the
H5Rdereference2 symbol is missing from the rhdf5.so library. This is
because the GNU linker, when linking with a static library, by default
only includes those objects from the static lib that are actually used
in the final output.